requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.1.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
blis==1.2.0
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpathlib==0.20.0
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.81.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.2
google-cloud-dlp==3.27.0
google-cloud-documentai==3.2.0
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.8.0
kubernetes==32.0.1
langchain==0.3.19
langchain-chroma==0.2.2
langchain-community==0.3.18
langchain-core==0.3.37
langchain-google-vertexai==2.0.13
langchain-huggingface==0.1.2
langchain-ollama==0.2.3
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.10
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
ml-dtypes==0.3.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.16.3
openai==1.64.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pillow==11.1.0
posthog==3.15.0
preshed==3.0.9
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pypdf==5.3.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
slack_sdk==3.34.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
starlette==0.45.3
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.2.2
tqdm==4.67.1
transformers==4.49.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.4
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
